CHROMIUM: pstore: fix dmi table to match all chrome systems
Turns out that DMI_SYSTEM_VENDOR is actually the native vendor. I
tested the original patch on a link that -- surprise, has Google as
vendor. *facepalm*. My bad when originally doing this for upstream.

So, this needs to be fixed upstream as well.

The only other data I can think of to probe on is Google_* in the version
string.  Checking with our firmware team, all systems should have this
(and nothing else than Chrome hardware should have the coreboot + Google_*
combination to date.

So, we'll switch to this. For future platforms we are going to move to
using an ACPI device to configure this instead of a DMI table (yay!),
so longer-term that will sort itself out.

BUG=chromium:349122
TEST=run KernelErrorPaths

Change-Id: I6e8571a256d3e0ba44457bf6e263b9505ff0d0f2
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188798