CHROMIUM: video: tegra: Allocate framebuffer dynamically
On 4G nyan big, the framebuffer address beyonds 4G since it's
reserved at the end of the whole memory. This brings some problems
in fbcon/fbmem frameworks because they only support 32bit address.

The easy solution is to mangle the framebuffer address to make it
under 4G while another better solution is creating the fb dynamically
with the help of iommu.

So this patch does:

- remove the framebuffer reservation codes
- allocate the framebuffer dynamically

BUG=chrome-os-partner:25909
TEST=Boot 4G nyan big OK(along with other dependent patches),
     framebuffer console works

Change-Id: I6ff92a39ff1d8938746095091c9a277f935d2c60
Signed-off-by: Mark Zhang <markz@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/188863
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>