drm/i915: Always set fifo count to zero in gen6_reset
There should not be a case where fifo count is other
than zero after a successful reset. Always set
count to zero, but be paranoid enough to warn.

The GT FIFO is bypassed when not in RC6 from both IA and SA. As we've
just reset the GPU and not yet enabled RC6, there is no way the FIFO
can be anything but 0. If it is non-zero, it's a HW bug, and we can
try to carry on by faking it. It should be noted that RC6 is highly
unlikely to work properly if this WARN fires, however the system
should continue on just fine.

v2: rebased

Suggested-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Mika Kuoppala <mika.kuoppala@intel.com>
[danvet: Extend the commit message as requested by Ben.]
Reviewed-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>