ACPI: Add sysfs "ddn" attribute to ACPI device objects
The information contained in _DDN (Dos Device Name) might be useful
for userspace utilities, so export it using a new sysfs "ddn"
attribute of ACPI device objects.

Note that the "ddn" attribute will be created as long as the _DDN
identifying object exists, regardless of whether or not its
evaluation is successful.

Signed-off-by: Lv Zheng <lv.zheng@intel.com>
[rjw: Changelog]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>