UPSTREAM: ASoC: dapm: Sprinkle lockdep asserts through the code
Try to spot locking issues by asserting that the DAPM mutex is held when
it should be. There's a bit of fun due to us not requiring the lock to be
held prior to the card being instantiated which mean we need to wrap the
assert in some paths and this isn't methodical by any stretch of the
imagination.

Signed-off-by: Mark Brown <broonie@linaro.org>
(cherry picked from commit f9fa2b1855a9ece2ec2a4f54b7f9131cd89257d6)
Signed-off-by: Kevin Strasser <kevin.strasser@intel.com>

BUG=chrome-os-partner:29649
TEST=With entire patch series applied, audio playback is working

Change-Id: I642997514ba202854b028279931a0d18a9ed0d9e
Reviewed-on: https://chromium-review.googlesource.com/215196
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Ben Zhang <benzh@chromium.org>
Tested-by: Ben Zhang <benzh@chromium.org>