arm, LLVMLinux: Remove use of named register for ARM percpu
The use of sp as a named register variable is not supported in clang and
the behavior they are working around in GCC may not even be present
in clang.

See commit 509eb76ebf97

Unfortunately, GCC doesn't treat a "memory" clobber on a non-volatile
asm block as a side-effect, and will happily re-order it before other
memory clobbers (including those in prempt_disable()) and cache the
value. This has been observed to break the cmpxchg logic in the slub
allocator, leading to livelock in kmem_cache_alloc in mainline kernels.

Because the GCC workaround causes numerous warnings, revert the change
until it can be verified that it is in fact an issue for clang.

Change-Id: I5778b240b1ba1ebd73fd9ce33ec5fff1dd8bbc39
Signed-off-by: Mark Charlebois <charlebm@gmail.com>
Signed-off-by: Behan Webster <behanw@converseincode.com>