BACKPORT: drm/tegra: dsi - Implement VDD supply support
The DSI controllers are powered by a (typically 1.2V) regulator. Usually
this is always on, so there was no need to support enabling or disabling
it thus far. But in order not to consume any power when DSI is inactive,
give the driver a chance to enable or disable the supply as needed.

Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 3b077afb3a7d84bffbf6598697a17655933671d2)
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>

BUG=chromium:385270
TEST=compiles

Change-Id: I6f941d712aa4ad0a7eda5acb7fc1a001afad1197
Reviewed-on: https://chromium-review.googlesource.com/208419
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>