BACKPORT: drm/tegra: dsi - Do not needlessly recompute pclk
In some cases the pixel clock used to not be correct, which is why it
had to be recomputed. It turns out that the reason why it wasn't correct
is that it was used wrongly. If used correctly there's not need for the
recomputation.

Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 76245adbc149170999340a61d171e29b86b02218)
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>

BUG=chromium:385270
TEST=compiles

Change-Id: I55d7d0bd2d367d4f84071ab6302bdb2ebf7e35fc
Reviewed-on: https://chromium-review.googlesource.com/208426
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>