CHROMIUM: video: tegra: sor: Do not attach again if attached
We have a check in sor attach function to see if the SOR is already
attached with dc. But we removed it in commit:

b7fe34b02322 CHROMIUM: video: tegra: Fix SOR attach sequence

This makes sometime(50%) the display doesn't come up on some boards.

So this patch adds the check back.

BUG=None
TEST=nyan & nyan big display is OK(tested 5 times on each board)
     LP0 stress test didn't find SOR timeouts(5 times, each ~200 cycles)

Change-Id: Ifacd456f20ccdf5efe072ef3307dbb8e0480fd18
Signed-off-by: Mark Zhang <markz@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/190448
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>