drm/anx7808: Increase delay at the end of power_off
This increases the delay at the end of the power_off routine to ensure
the value of the cable_det gpio is valid when we exit. The gpio bounces
as a result of powering down the chip, so waiting 25ms ensures that it's in
steady state.

BUG=chromium:351844
TEST=Tested on spring by entering/exiting docked mode repeatedly

Change-Id: Iad2efec2e8d4a8c722a8b7539fce006c3a9cb4dd
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/190494
Reviewed-by: David Riley <davidriley@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/194016
Reviewed-by: Jongpil Jung <jongpil19.jung@samsung.com>
Commit-Queue: Jongpil Jung <jongpil19.jung@samsung.com>
Tested-by: Jongpil Jung <jongpil19.jung@samsung.com>