CHROMIUM: ASoC: dma - Don't try to pause or resume if it isn't supported
If the audio device doesn't support pause or resume then it will not
be paused or resumed, through the ioctl, but pause and resume were
being called and assumed to succeed from the DMA engine.  Check that
they are supported before calling.  If pause isn't supported, instead
terminate all the transaction before suspending.  Since the driver
doesn't support resume, prepare and start will be called again after
the system resumes.

BUG=chrome-os-partner:26821
TEST=play audio across suspend/resume through cras.

Change-Id: I4109782f8d57a54f7a03c7ee92ebd0662f0267f3
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/190846
Reviewed-by: Stephen Warren <swarren@nvidia.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>