mfd: devicetree: bindings: add pm8xxx RTC description
The PM8xxx family of PMICs contain an RTC.  This RTC is described as a
subnode of the PM8xxx.  Document these bindings, and replace the pwrkey
node in the example with the RTC, which is now described in this document.

While we're here, add a short description to the device tree bindings
describing what the the PM8xxx devices are and how they are expected to be
used.

Signed-off-by: Josh Cartwright <joshc@codeaurora.org>
Cc: Stephen Boyd <sboyd@codeaurora.org>
Cc: Alessandro Zummo <a.zummo@towertech.it>
Cc: Rob Herring <robherring2@gmail.com>
Cc: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>