CHROMIUM: ARM: tegra124: dvfs: G CPU DVFS table update
Add one more entry for 204MHz and adjust some of the safe voltage for
some low frequencies, min_millivolts is reduced as well since we are
going to use DFLL as G CPU parent for frequency above 204MHz. The DFLL
CVB table needs to be updated as well to work with this changes.

CQ-DEPEND=CL:191209
BUG=None
TEST=Merge the DFLL DTS table change and then check the CPU G parent,
     it should be DFLL for frequency higher or equeal to 204MHz, test
     suspend stress to see if it works.

Change-Id: I5a8c3ff549550e247742be88401de4637d2bd9ff
Signed-off-by: Bill Huang <bilhuang@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/191208
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>