CHROMIUM: drm/i915: do not explicitly disable backlight in panel_off
Per eDP spec, we must disable the backlight in order to power down the
panel. However, in our code, we have always disabled the backlight
before we try to turn off the panel. The assertions from the previous
patch make sure this is the case.

Signed-off-by: Jesse Barnes <jbarnes@virtuousgeek.org>
Signed-off-by: Wayne Boyer <wayne.boyer@intel.com>

BUG=chrome-os-partner:25159
TEST=suspend/resume test on instrumented system shows correct signaling
on oscope.

Change-Id: I68ae4ee80cee40aa4745697ce7acf3883ffb721d
Reviewed-on: https://chromium-review.googlesource.com/194250
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Tested-by: Wayne Boyer <wayne.boyer@intel.com>
Commit-Queue: Wayne Boyer <wayne.boyer@intel.com>