CHROMIUM: mmc: tegra: reintroduce shutdown callback
Commit 1d4b798 ["CHROMIUM: mmc: tegra: power cycle card at boot"]
removed the shutdown callback to disable the power GPIO in favor of
toggling it at boot.  Some SD cards, however, still fail to get
detected at boot with this change.  Add back the shutdown callback
so that these cards can get detected reliably, but leave the power
cycling at boot in case the previous shutdown was not clean.

BUG=chrome-os-partner:27053
TEST=SD card that was previously not getting detected now works.

Change-Id: If5caa297bbfdefe5e878af49a295f055b96b68aa
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/192161
Reviewed-by: Dylan Reid <dgreid@chromium.org>