CHROMIUM: config: enable UHID
UHID provides a mechanism for userspace to implement HID I/O Drivers,
performing the raw data I/O to the low level device and pushing HID
Reports to the kernel for parsing.

It's required for HID over GATT (HID on Bluetooth Low Energy) because,
unlike HID on classic Bluetooth, the GATT protocol is implemented in
userspace whereas the L2CAP protocol on which HID is based is
implemented in the kernel.

BUG=chromium:345982
TEST=`emerge-$BOARD chromeos-kernel-3_10` and check for /dev/uhid

Change-Id: Id9039657d1f9137e770ac30a3baef3ce512ec483
Signed-off-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/192381
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>