CHROMIUM: cros_ec: Add host commands to cros_ec_accel driver
Added ability to modify accelerometer parameters to the cros_ec_accel driver
by sending host commands to EC. This adds four new sysfs files:
sampling_frequency for the rate at which the EC samples all motion sensors (in ms),
peak_scale is the max range setting of the sensor (in +/- G's), frequency
is the setting for output data rate (ODR) of the accelerometer, and lid angle
offset is the lid angle threshold for the EC to be in tablet mode in which
the keyboard is not used as a wake source in S3.

BUG=chrome-os-partner:27321
TEST=manual
On chromebook with accelerometer, tested that I can read/write to sampling
frequency, accelerometer frequency (ODR), and peak scale (sensor range).
Each of these require a host command to the EC. Note that when writing to
the peak scale and frequency, the EC rounds the input arg up to the nearest
valid value.

$ cd /sys/bus/iio/devices/iio:device0
$ cat in_accel_lid_sampling_frequency
10
$ echo "150" > in_accel_lid_sampling_frequency
$ cat in_accel_lid_sampling_frequency
150

$ cat in_accel_lid_peak_scale
2
$ echo "5" > in_accel_lid_peak_scale
$ cat in_accel_lid_peak_scale
8

$ cat in_accel_lid_frequency
100000
$ echo "25000" > in_accel_lid_frequency
$ cat in_accel_lid_frequency
50000

$ cat in_angl_offset
180
$ echo "210" > in_angl_offset
$ cat in_angl_offset
210

Change-Id: Ic6b384e9bf3f974c854edb094c1f9a7cd9a26763
Signed-off-by: Alec Berg <alecaberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/192722
Reviewed-by: Olof Johansson <olofj@chromium.org>