drm/i915/vlv: assert and de-assert sideband reset on resume
This is a bit like the CMN reset de-assert we do in DPIO_CTL, except
that it resets the whole common lane section of the PHY.  This is
required on machines where the BIOS doesn't do this for us on resume to
properly re-calibrate and get the PHY ready to transmit data.

Without this patch, such machines won't resume correctly much of the time,
with the symptom being a 'port ready' timeout and/or a link training
failure.

BUG=chrome-os-partner:25517
TEST=by hand on glimmer: suspend/resume a few times, with and without HDMI plugged in

Change-Id: I3ee3e683db7638ee06ede982f957ebd3d2dfe934
Signed-off-by: Jesse Barnes <jesse.barnes@intel.com>
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/194325
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>