UPSTREAM: MIPS: mark R4K clockevent device with CLOCK_EVT_FEAT_PERCPU
The CLOCK_EVT_FEAT_PERCPU flag indicates that a clockevent device is
only configurable by the CPU for which it is registered, and thus cannot
be used as the tick broadcast device. That property is true of the R4K
timer, which is inaccessible from other cores.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
(cherry picked from commit d8107efd8a3c15ec5885dbe1d3168e26c6b3e2e4)

BUG=None
TEST=Build and boot on Malta

Change-Id: I9ac14d1b66a90b1d44fbcc2a026c0e52b35039a7
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/208701
Reviewed-by: Olof Johansson <olofj@chromium.org>