CHROMIUM: ARM: dts: nyan: add vmmc supply for SD card
The power-gpio field, which should be used for power/reset toggles
for wifi, was being used to turn on/off the SD card power supply.
This patch turns the card power supply into a fixed regulator and
assigns it as the vmmc-supply. The mmc subsystem will handle powering
it on and off.

BUG=chrome-os-partner:27053
TEST=Reboot cycles

Change-Id: I94dceb8b50f14bb8b090c477b702ff32bd10e29c
Signed-off-by: Christopher Freeman <cfreeman@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/195107
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>