UPSTREAM: drm/i915: Remove vblank wait from haswell_write_eld
The pipe is off at that point in time, so a vblank wait is simply a
50ms wait. Caught by Jesse's verbose "make vblank wait timeouts WARN"
patch. We've probably had a few versions of this float around already.

To document assumptions put a pipe assert into the same place. And
also add a posting read.

If we ever decide to update the eld and infoframes while the pipe is
already on (e.g. for fastboot) then there's lots of work to do. So
better properly document all the hidden assumptions.

Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit c79057922ed6c2c6df1214e6ab4414fea1b23db2)
Signed-off-by: Gary Wang <gary.c.wang@intel.com>

BUG=chrome-os-partner:35745
TEST=Autotest power_Resume.control.resume_constraint

Change-Id: I9db60e5b49cc6e9a7062bf51f5385c2f6f628277
Signed-off-by: Gary Wang <gary.c.wang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/246433
Reviewed-by: Shan-fu Chiou <shan-fu.chiou@intel.com>
Tested-by: Ryan Lin <ryan.lin@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>