FROMLIST: dmaengine: qcom_bam_dma: Add v1.3.0 driver support
This patch adds support for the v1.3.0 version of the BAM dma ip block.  This
patch adds register access abstraction to deal with the changes to the register
map between the two versions.  Blocks of registers moved around within the
address space, and multipliers used for calculating the pipe registers changed
as well.

Signed-off-by: Andy Gross <agross@codeaurora.org>
BUG=chrome-os-partner:31587
TEST=build/boot on AP-148 and storm P0

Change-Id: I65115f9cf1915b5352ffa52081627cfdacf84e54
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/215119