CHROMIUM: clk: tegra: remove EMC resume support
This was intended to sync up the kernel's state with respect to the EMC
clock after resume since the boot ROM may restore the EMC clock at a
different rate and/or with a different parent.  It's unsafe, however,
to reparent a clock at syscore time since it calls into the filesystem
layer.  Let the EMC driver lock the frequency across suspend/resume
instead.

BUG=chrome-os-partner:28140
TEST=suspend_stress_test on Big; no more spinlock BUGs.

Change-Id: I5b148b9efd3f03865d4460768eb863c07aab9571
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/196138
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Bill Huang <bilhuang@nvidia.com>
Reviewed-by: Danny Huang <dahuang@nvidia.com>
Tested-by: Joseph Lo <josephl@nvidia.com>