CHROMIUM: drm/i915: Provide valleyview backlight fallback value
The fallback values for VLV reflect the Rambi3 panel values.

This patch introduces min_brightness member, which defined a part of VBT
that has some confusion. The field itself is a byte ranging from 0-255.
How this value is supposed to be used by the driver is unclear. Though
no code uses it yet, I've written a comment explaining that I intend to
use the number in there as an absolute value that represents the minimum
duty cycle ratio of the panel.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Wayne Boyer <wayne.boyer@intel.com>

BUG=chrome-os-partner:25159
TEST=suspend/resume test on instrumented system shows correct
signaling on oscope.  Manual brightness checks.

Change-Id: I9f0d8d4fe2778c4768b06459b0a83293b1c6f023
Reviewed-on: https://chromium-review.googlesource.com/196607
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Tested-by: Wayne Boyer <wayne.boyer@intel.com>
Commit-Queue: Wayne Boyer <wayne.boyer@intel.com>