CHROMIUM: drm/i915/vlv: Scale backlight to min duty ratio
On VLV specifically, going too low runs the risk of getting the BLC_EN
signal out of sync, preventing resume from working correctly.  Scale
/sys/class/backlight at this level to prevent userspace from doing this
on suspend.

This gets rid of the explicitly hardcoded value for a previous/similar
patch. Instead relies on the VBT to provide the value, and if VBT does
not provide one, then it uses a hardcoded 5% min duty cycle. This was
defined a couple patches ago for the same fallback on VLV.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Wayne Boyer <wayne.boyer@intel.com>

BUG=chrome-os-partner:25159
TEST=suspend/resume test on instrumented system shows correct
signaling on oscope.  Manual brightness checks.

Change-Id: Ibbbc33338ef1659d0f797d6a33d3727a45ad95d5
Reviewed-on: https://chromium-review.googlesource.com/196609
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Tested-by: Wayne Boyer <wayne.boyer@intel.com>
Commit-Queue: Wayne Boyer <wayne.boyer@intel.com>